+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## xxx
 
 * Change `init()` signature to named parameters and added optional `version`.
+* Added `setTags()` and `setCustomData()` methods.
+* Added `tags` and `customData` to the error sending methods.
+* Error sending methods can generate a stacktrace automatically.
+* Added `setUser()` method with `RaygunUserInfo`.
 
 ## 0.1.0